requirements.txt
# ─── MCP Server Dependencies ───
mcp[cli]>=1.0.0
httpx>=0.27.0
pytrends>=4.9.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# ─── Backend Dependencies ───
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.0.0

# ─── Auth (when ready) ───
# python-jose[cryptography]
# passlib[bcrypt]

# ─── Frontend Dependencies (Phase 3) ───
# (managed separately via npm/package.json)